# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Nexus AI Business Orchestration API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /actions/{id}/undo:
    post:
      operationId: undoAction
      tags: [actions]
      summary: Revert a recorded agent action
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Reverted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  reverted: { type: string }
...
  /agents/{agentId}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get a specific agent with task details
      parameters:
        - name: agentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
...
  /chat/messages:
    get:
      operationId: listChatMessages
      tags: [chat]
      summary: List orchestrator chat history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the CEO Orchestrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Message sent and orchestrator response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  # ─── IDEAS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Agent:
      type: object
      required: [id, name, department, status, activeTasks]
      properties:
        id:
          type: string
        name:
          type: string
        department:
          type: string
          description: "orchestrator | marketing | sales | research | finance"
        status:
          type: string
          description: "active | idle | busy | standby"
        activeTasks:
          type: integer
        currentTask:
          type: ["string", "null"]
        lastActive:
          type: ["string", "null"]
        description:
          type: string
        capabilities:
          type: array
          items:
            type: string
...
    ChatMessage:
      type: object
      required: [id, role, content, timestamp]
      properties:
        id:
          type: integer
        role:
          type: string
          description: "user | orchestrator | system"
        content:
          type: string
        agentActions:
          type: ["string", "null"]
          description: JSON array of agent actions taken
        timestamp:
          type: string
...
    ChatMessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        executionLevel:
          type: string
          enum: [green, amber, red]
          description: green = act silently, amber = act and report, red = propose only
    ChatResponse:
      type: object
      required: [userMessage, orchestratorResponse]
      properties:
        userMessage:
          $ref: "#/components/schemas/ChatMessage"
        orchestratorResponse:
          $ref: "#/components/schemas/ChatMessage"
        agentActions:
          type: array
          items:
            $ref: "#/components/schemas/AgentAction"
    AgentAction:
      type: object
      required: [agentId, agentName, action]
      properties:
        agentId:
          type: string
        agentName:
          type: string
        action:
          type: string
        result:
          type: ["string", "null"]

    # ─── IDEA SCHEMAS ─────────────────────────────────────────────────────────
...